refactor: Split breadcrumb files
