+++ src/drive/web/modules/navigation/Breadcrumb/PreviousButton.jsx
@@ -1,0 +1,8 @@
+import React from 'react'
+import styles from './breadcrumb.styl'
+
+export const PreviousButton = ({ onClick }) => (
+  <button className={styles['fil-path-previous']} onClick={onClick} />
+)
+
+export default PreviousButton
