+++ src/drive/web/modules/navigation/Breadcrumb/breadcrumb.styl
@@ -2,7 +2,7 @@
 @require 'settings/palette.styl'
 @require 'settings/breakpoints.styl'
 @require 'settings/z-index.styl'
-@require '../../../styles/coz-bar-size.styl'
+@require '../../../../styles/coz-bar-size.styl'
 
 .fil-path-backdrop
   flex 1 1 auto
@@ -31,7 +31,7 @@
     height     $coz-bar-size
     border     0
     flex-shrink 0
-    background embedurl('../../../assets/icons/icon-arrow-left.svg') center center no-repeat
+    background embedurl('../../../../assets/icons/icon-arrow-left.svg') center center no-repeat
 
 .fil-path-down
     display     none
@@ -39,7 +39,7 @@
     height      .625rem
     margin-left .4375rem
     border      0
-    background  embedurl('../../../assets/icons/icon-arrow-down.svg') center center no-repeat
+    background  embedurl('../../../../assets/icons/icon-arrow-down.svg') center center no-repeat
 
 .fil-path-current-name
     text-overflow  ellipsis
