+++ src/drive/web/modules/navigation/Breadcrumb/getFolderPath.js
